+++ src/nes/mappers/sunsoft_fme7.hpp
@@ -16,7 +16,7 @@ public:
     uint16_t getPRGPageSize() override {
         return 0x2000;
     }
-    uint16_t getSRAMSize() override {
+    uint32_t getSRAMSize() override {
         return 0x8000;
     }
